+++ src/clientLoader.js
@@ -1,4 +1,6 @@
 
+// polyfills
+import 'babel-polyfill';
 import main from './client';
 
 function run() {
@@ -11,17 +13,21 @@ function run() {
 }
 
 if (!global.Intl) {
+  // You can show loading banner here
+
   require.ensure([
+    // Add all large polyfills here
     'intl',
     'intl/locale-data/jsonp/en.js',
     'intl/locale-data/jsonp/cs.js',
   ], require => {
+    // and require them here
     require('intl');
     require('intl/locale-data/jsonp/en.js');
     require('intl/locale-data/jsonp/cs.js');
 
     run();
-  }, 'intl');
+  }, 'polyfills');
 } else {
   run();
 }