+++ src/components/Link/Link.js
@@ -8,8 +8,6 @@
  */
 
 import React, { Component, PropTypes } from 'react';
-import { connect } from 'react-redux';
-import { navigate } from '../../actions/route';
 
 function isLeftClickEvent(event) {
   return event.button === 0;
@@ -22,21 +20,16 @@ function isModifiedEvent(event) {
 class Link extends Component {
 
   static propTypes = {
-    to: PropTypes.oneOfType([PropTypes.string, PropTypes.object]).isRequired,
+    to: PropTypes.string.isRequired,
     children: PropTypes.node,
     onClick: PropTypes.func,
-
-    // actions
-    navigate: PropTypes.func,
   };
 
   static contextTypes = {
-    createHref: PropTypes.func.isRequired,
+    history: PropTypes.object.isRequired,
   };
 
   handleClick = (event) => {
-    let allowTransition = true;
-
     if (this.props.onClick) {
       this.props.onClick(event);
     }
@@ -46,43 +39,18 @@ class Link extends Component {
     }
 
     if (event.defaultPrevented === true) {
-      allowTransition = false;
+      return;
     }
 
     event.preventDefault();
-
-    if (allowTransition) {
-      if (this.props.to) {
-        this.props.navigate(this.props.to);
-      } else {
-        this.props.navigate({
-          pathname: event.currentTarget.pathname,
-          search: event.currentTarget.search,
-        });
-      }
-    }
+    this.context.history.push(this.props.to);
   };
 
   render() {
-    const {
-      to,
-      navigate, // eslint-disable-line no-unused-vars, no-shadow
-      children,
-      ...props,
-    } = this.props;
-    return (
-      <a href={this.context.createHref(to)} {...props} onClick={this.handleClick}>
-        {children}
-      </a>
-    );
+    const { to, children, ...props } = this.props;
+    return <a href={to} {...props} onClick={this.handleClick}>{children}</a>;
   }
 
 }
 
-const mapState = null;
-
-const mapDispatch = {
-  navigate,
-};
-
-export default connect(mapState, mapDispatch)(Link);
+export default Link;