+++ src/config.js
@@ -13,7 +13,7 @@ export const port = process.env.PORT || 3000;
 export const host = process.env.WEBSITE_HOSTNAME || `localhost:${port}`;
 
 // default locale is the first one
-export const locales = ['en', 'cs'];
+export const locales = ['en-US', 'cs-CZ'];
 
 export const databaseUrl = process.env.DATABASE_URL || 'sqlite:database.sqlite';