+++ src/routes/content/index.js
@@ -30,7 +30,10 @@ export default {
     if (resp.status !== 200) throw new Error(resp.statusText);
     const { data } = await resp.json();
     if (!data || !data.content) return undefined;
-    return <Content {...data.content} />;
+    return {
+      title: data.content.title,
+      component: <Content {...data.content} />,
+    };
   },
 
 };